HHH-16009 Code refactoring
